refactor to centralize initialization of sessions, remove os dependency and fix docker file system references
