+++ app/__init__.py
@@ -2,129 +2,17 @@ from flask import *
 from flask_cors import CORS
 from flask_socketio import SocketIO
 from pathlib import Path
-import shutil
 import yaml
 import os
 
-APP_FOLDER = 'app'
+from .main.model import PicoFermSession, PicoBrewSession
+from .main.config import brew_active_sessions_path
+from .main.session_parser import restore_active_sessions, active_brew_sessions, active_ferm_sessions
+from .main.routes_frontend import initialize_data
 
-if os.getenv("BASE_PATH"):
-    BASE_PATH = Path(os.getenv("BASE_PATH"))
-else:
-    BASE_PATH = Path(__file__).parents[1]
-
-# recipe paths
-ZYMATIC_RECIPE_PATH = str(BASE_PATH.joinpath('app/recipes/zymatic'))
-ZSERIES_RECIPE_PATH = str(BASE_PATH.joinpath('app/recipes/zseries'))
-PICO_RECIPE_PATH = str(BASE_PATH.joinpath('app/recipes/pico'))
-
-# sessions paths
-BREW_ACTIVE_PATH = str(BASE_PATH.joinpath('app/sessions/brew/active'))
-BREW_ARCHIVE_PATH = str(BASE_PATH.joinpath('app/sessions/brew/archive'))
-FERM_ACTIVE_PATH = str(BASE_PATH.joinpath('app/sessions/ferm/active'))
-FERM_ARCHIVE_PATH = str(BASE_PATH.joinpath('app/sessions/ferm/archive'))
-
-ZYMATIC_LOCATION = {
-    'PassThru': '0',
-    'Mash': '1',
-    'Adjunct1': '2',
-    'Adjunct2': '3',
-    'Adjunct3': '4',
-    'Adjunct4': '5',
-    'Pause': '6',
-}
-
-ZSERIES_LOCATION = {
-    'PassThru': '0',
-    'Mash': '1',
-    'Adjunct1': '2',
-    'Adjunct2': '3',
-    'Adjunct3': '4',
-    'Adjunct4': '5',
-    'Pause': '6',
-}
-
-PICO_LOCATION = {
-    'Prime': '0',
-    'Mash': '1',
-    'PassThru': '2',
-    'Adjunct1': '3',
-    'Adjunct2': '4',
-    'Adjunct3': '6',
-    'Adjunct4': '5',
-}
-PICO_SESSION = {
-    0: 'Brewing',
-    1: 'Deep Clean',
-    2: 'Sous Vide',
-    4: 'Cold Brew',
-    5: 'Manual Brew',
-}
-
-
-active_brew_sessions = {}
-
-
-class PicoBrewSession():
-    def __init__(self):
-        self.file = None
-        self.filepath = None
-        self.alias = ''
-        self.created_at = None
-        self.name = 'Waiting To Brew'
-        self.type = 0
-        self.step = ''
-        self.session = ''   # session guid
-        self.id = -1        # session id (integer)
-        self.recovery = ''
-        self.remaining_time = None
-        self.is_pico = True
-        self.data = []
-
-    def cleanup(self):
-        if self.file and self.filepath:
-            self.file.close()
-            shutil.move(str(self.filepath), str(Path(BREW_ARCHIVE_PATH)))
-        self.file = None
-        self.filepath = None
-        self.created_at = None
-        self.name = 'Waiting To Brew'
-        self.type = 0
-        self.step = ''
-        self.session = ''
-        self.id = -1
-        self.recovery = ''
-        self.remaining_time = None
-        self.data = []
-
-
-active_ferm_sessions = {}
-
+BASE_PATH = Path(__file__).parents[1]
 
 server_cfg = {}
-
-
-class PicoFermSession():
-    def __init__(self):
-        self.file = None
-        self.filepath = None
-        self.alias = ''
-        self.uninit = True
-        self.voltage = '-'
-        self.start_time = None
-        self.data = []
-
-    def cleanup(self):
-        if self.file and self.filepath:
-            self.file.close()
-            shutil.move(str(self.filepath), str(Path(FERM_ARCHIVE_PATH)))
-        self.file = None
-        self.filepath = None
-        self.uninit = True
-        self.voltage = '-'
-        self.start_time = None
-        self.data = []
-
 
 socketio = SocketIO()
 
@@ -133,16 +21,27 @@ def create_app(debug=False):
     """Create an application."""
     app = Flask(__name__)
     CORS(app)
-    app.config.update(SECRET_KEY='bosco', CORS_HEADERS='Content-Type')
 
     from .main import main as main_blueprint
 
     # ----- Routes ----------
     app.register_blueprint(main_blueprint)
     socketio.init_app(app)
-    cfg_file = BASE_PATH.joinpath('app/config.yaml') if os.getenv("BASE_PATH") else BASE_PATH.joinpath('config.yaml')
+
+    cfg_file = BASE_PATH.joinpath('config.yaml')
     with open(cfg_file, 'r') as f:
         server_cfg = yaml.safe_load(f)
+
+    app.config.update(
+        SECRET_KEY='bosco', 
+        CORS_HEADERS='Content-Type',
+        RECIPES_PATH=BASE_PATH.joinpath('app/recipes'),
+        SESSIONS_PATH=BASE_PATH.joinpath('app/sessions'),
+    )
+
+    with app.app_context():
+        restore_active_sessions()
+        initialize_data()
 
     if 'aliases' in server_cfg:
         machine_types = ["ZSeries", "Zymatic", "PicoBrew", "PicoFerm"]
