+++ docker-compose.yml
@@ -6,14 +6,13 @@ services:
       dockerfile: Dockerfile
     environment: 
       FLASK_ENV: development
-      BASE_PATH: /
     volumes:
       - type: bind
         source: ./app/recipes
-        target: /app/recipes
+        target: /picobrew/app/recipes
       - type: bind
         source: ./app/sessions
-        target: /app/sessions
+        target: /picobrew/app/sessions
     ports:
       - 80:80
